ExConvCode: fix unaligned read
The ExConvCode implementation used the _mm_load_ps intrinsic
to load 4 f32 into an __m128. However, when accumulute is
called from dualEncode<block, u8>, the input is not aligned,
leading to a segfault.
Instead, we use the _mm_loadu_ps intrinsic which allows
unaligned reads. The non-sse fallback is implemented using
a memcpy, as it would otherwise also result in an unaligned
read.